Refactor background job manager into dedicated module
